\name{cnO2}
\docType{data}
\alias{cnO2}
\title{Catalogue of Coherent Networks of Order 2}
\description{
  This data sets provides a catalogue of the network graph, signature and minimal cut-sets of all coherent networks of node order 2.
}
\usage{data(cnO2)}
\format{A list object, one item for each such network.  Each item is itself a list, with the elements \code{$graph}, \code{$cutsets} and \code{$signature}.}
\source{Derived in the thesis Aslett (2012).}
\references{
Aslett, L. J. M. (2012), \emph{MCMC for Inference on Phase-type and Masked System Lifetime Models}, PhD Thesis, Trinity College Dublin.
}
\keyword{datasets}
