\name{ReliabilityTheory-package}
\alias{ReliabilityTheory-package}
\alias{ReliabilityTheory}
\docType{package}
\title{Structural Reliability Theory Toolbox}
\description{
A collection of tools for working structural reliability problems, such as catalogues of system signatures and Bayesian inferential functions.
}
\details{
\tabular{ll}{
Package: \tab ReliabilityTheory\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2012-10-29\cr
License: \tab GPL-2 | GPL-3\cr
LazyLoad: \tab yes\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Louis J. M. Aslett, \email{aslett@stats.ox.ac.uk} (\url{http://www.louisaslett.com})
}
\references{
Aslett, L. J. M. (2012), \emph{MCMC for Inference on Phase-type and Masked System Lifetime Models}, PhD Thesis, Trinity College Dublin.

Samaniego, F. J. (2007), \emph{System Signatures and Their Applications in Engineering Reliability}, Springer.
}
\keyword{ reliability theory }
\keyword{ system signature }
\keyword{ signature }
%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{}
