\name{sccsO3}
\docType{data}
\alias{sccsO3}
\title{Catalogue of Simply Connected Coherent Systems of Order 3}
\description{
  This data set provides a catalogue of the network graph, signature and minimal cut-sets of all simply connected coherent systems of order 3.
}
\usage{data(sccsO3)}
\format{A list object, one item for each such systems.  Each item is itself a list, with the elements \code{$graph}, \code{$cutsets} and \code{$signature}.}
\source{Derived in the thesis Aslett (2012).}
\references{
Aslett, L. J. M. (2012), \emph{MCMC for Inference on Phase-type and Masked System Lifetime Models}, PhD Thesis, Trinity College Dublin.
}
\keyword{datasets}
